Safari-609.1.11.1/LayoutTests/fast/html/menuitem-element.html

<!DOCTYPE html><!-- webkit-test-runner [ enableMenuItemElement=true ] -->
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body>
<script>
description('Various tests for the menuitem element.');

debug('MenuItem is recognized:')
shouldBeTrue('document.createElement("menuitem") instanceof HTMLMenuItemElement');

</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
