LayoutTests/fast/html/menuitem-element-expected.txt

Various tests for the menuitem element.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


MenuItem is recognized:
PASS document.createElement("menuitem") instanceof HTMLMenuItemElement is true
PASS successfullyParsed is true

TEST COMPLETE
